need to test

+++ client.c
@@ -58,7 +58,7 @@ void Client_sendMessage(char* messageTx, int size)
         exit(EXIT_FAILURE);
     }
 }
-char* Client_getMessage(int size)
+char* Client_getMessage(long size)
 {
     // first message from
     char* messageRx = malloc(sizeof(char) * size);
@@ -70,10 +70,10 @@ char* Client_getMessage(int size)
     return messageRx;
 }
 
-int Client_getMessageSize(void)
+long Client_getMessageSize(void)
 {
-    int size = 0;
-    if (recv(s_options.socket_fd, &size, sizeof(int), 0) < 0) {
+    long size = 0;
+    if (recv(s_options.socket_fd, &size, sizeof(long), 0) < 0) {
         perror("Error while receiving server's msg\n");
         exit(EXIT_FAILURE);
     }
